+++ src/index.html
@@ -31,6 +31,8 @@
 
     <include src="./html/footer.html"></include>
 
+    <include src="./html/team-window.html"></include>
+
     <include src="./html/modal-window.html"></include>
     <include src="./html/card-film.html"></include>
   </body>
