+++ src/index.js
@@ -1,1 +1,2 @@
 // import './js/***.js';
+import './js/team-window.js'
