+++ src/sass/layout/_all-partials.scss
@@ -3,3 +3,4 @@
 @import './footer';
 @import './modal-window';
 @import './pagination';
+@import './team-window';
